fix: use bigint-compatible business code counter

+++ src/Auran.Clinic.Domain/Entities/CodeCounter.cs
@@ -10,5 +10,5 @@ public class CodeCounter : BaseEntity
     public CodeType CodeType { get; set; }
     public required string Prefix { get; set; }
     public int Year { get; set; }
-    public int LastNumber { get; set; }
+    public long LastNumber { get; set; }
 }
